% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{COMPLETERS_SPEC}
\alias{COMPLETERS_SPEC}
\title{Layout information about the tables that report completion data}
\format{
\subsection{\code{COMPLETERS_SPEC}}{

A list with three elements.
\subsection{A}{

\strong{Completion by major}

This is a huge file because of the highly detailed breakdown of majors.
}

\subsection{B}{

\strong{Completion by Population and Gender}

This table reports graduates in combinations of sex and race/ethnicity
}

\subsection{C}{

\strong{Completion by Population}

This table reports graduates by single population, like sex, race/ethnicity, or age
}

}
}
\source{
\url{https://nces.ed.gov/ipeds/datacenter/DataFiles.aspx}
}
\usage{
COMPLETERS_SPEC
}
\description{
Layout information about the tables that report completion data
}
\keyword{datasets}
