% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CORNELLS}
\alias{CORNELLS}
\title{IPEDS data from 2022 for schools with "Cornell" in their names.}
\format{
\subsection{\code{CORNELLS}}{

A list with two elements.
\subsection{Directory}{

An object of class \code{spec_tbl_df}
(inherits from \code{tbl_df}, \code{tbl}, \code{data.frame})
with 3 rows and 73 columns.
Detailed characteristics of each institution in 2022.
}

\subsection{Completers}{

An object of class \code{spec_tbl_df}
(inherits from \code{tbl_df}, \code{tbl}, \code{data.frame})
with 7 rows and 36 columns.
Graduates in 2022 by population from each institution.
}

\subsection{Outcomes}{

An object of class \code{spec_tbl_df}
(inherits from \code{tbl_df}, \code{tbl}, \code{data.frame})
with 20 rows and 54 columns.
Overall 4-, 6-, and 8-year completion from each institution
}

}
}
\source{
\url{https://nces.ed.gov/ipeds/datacenter/DataFiles.aspx}
}
\usage{
CORNELLS
}
\description{
IPEDS data from 2022 for schools with "Cornell" in their names.
}
\keyword{datasets}
