% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_ipeds_table.R
\name{fetch_ipeds_table}
\alias{fetch_ipeds_table}
\title{Download one data table from IPEDS}
\usage{
fetch_ipeds_table(.year, .prefix, .suffix, .destination_path)
}
\arguments{
\item{.year}{\verb{<int>} the year that the table describes}

\item{.prefix}{\verb{<chr>} a prefix that describes the type of data, like "GR" or "EFFY"}

\item{.suffix}{\verb{<chr>} the rest of the file name, like ".zip" or "_PELL_SSL.zip"}

\item{.destination_path}{\verb{<chr>} the folder where the file should be saved}
}
\value{
\verb{<str>} the full path to the saved file
}
\description{
Download one data table from IPEDS
}
