% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GLOSSARIES_FOR_EFFY}
\alias{GLOSSARIES_FOR_EFFY}
\title{Mapping details about enrollment from a complex field in the table of head counts.}
\format{
\subsection{\code{GLOSSARIES_FOR_EFFY}}{

A list of five data frames.
\subsection{EFFYALEV}{

An object of class \code{spec_tbl_df}
(inherits from \code{tbl_df}, \code{tbl}, \code{data.frame})
with 27 rows and 4 columns.

\describe{
\item{EFFYALEV}{\verb{<int>} a code that maps to the rest of the columns}
\item{Time}{\verb{<chr>} whether full-time, part-time, or all students}
\item{Level}{\verb{<chr>} undergraduate, graduate, or all students}
\item{Goal}{\verb{<chr>} degree-seeking, non-degree-seeking, or all}
\item{Enrollment}{\verb{<chr>} first-time, transfer, continuing, or all students}
}
}

\subsection{EFFYLEV}{

An object of class \code{spec_tbl_df}
(inherits from \code{tbl_df}, \code{tbl}, \code{data.frame})
with 4 rows and 2 columns.

\describe{
\item{EFFYLEV}{\verb{<int>} a code that maps to the rest of the columns}
\item{Level of Study}{\verb{<chr>} All, Graduate, Undergraduate, or Detail}
}
}

\subsection{LSTUDY}{

An object of class \code{spec_tbl_df}
(inherits from \code{tbl_df}, \code{tbl}, \code{data.frame})
with 3 rows and 2 columns.

\describe{
\item{LSTUDY}{\verb{<int>} a code that maps to the rest of the columns}
\item{Original Level of Study}{\verb{<chr>} TOtal, Graduate, or Undergraduate}
}
}

\subsection{EFY}{

An object of class \code{spec_tbl_df}
(inherits from \code{tbl_df}, \code{tbl}, \code{data.frame})
with 34 rows and 4 columns.

\describe{
\item{EFY}{\verb{<int>} a 7- or 7-letter code that maps to the rest of the columns}
\item{Population}{\verb{<chr>} A population name similar to those from OMB97}
\item{Sex}{\verb{<chr>} Female, male, all, or another gender identity}
\item{Description}{\verb{<chr>} The population description from the dictionary file}
}
}

}
}
\source{
\url{https://nces.ed.gov/ipeds/datacenter/DataFiles.aspx}
}
\usage{
GLOSSARIES_FOR_EFFY
}
\description{
Mapping details about enrollment from a complex field in the table of head counts.
}
\keyword{datasets}
