% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GLOSSARIES_FOR_GR_TYPES}
\alias{GLOSSARIES_FOR_GR_TYPES}
\title{Mapping details about credentials and success from a complex field in the table of graduation rates.}
\format{
\subsection{\code{GLOSSARIES_FOR_GR_TYPES}}{

A list of two data frames.
\subsection{Cohort}{

An object of class \code{spec_tbl_df}
(inherits from \code{tbl_df}, \code{tbl}, \code{data.frame})
with 28 rows and 4 columns.

\describe{
\item{GR_TYPE}{\verb{<int>} a code that maps to the rest of the columns}
\item{Years}{\verb{<int>} Whether the cohort comes from a 2- or 4-year institution.}
\item{Award Sought}{\verb{<fct>}
Cohorts are defined by their institution and the type of award their students enrolled to earn.
Values include "Certificate," "Associate's," "Bachelor's," or "Any." }
\item{Statistic}{\verb{<fct>}
The quantity reported in rows with this GR_TYPE code.
Possible values are:
"Everyone",
"Excluded from cohort",
"Adjusted cohort",
"Transferred out",
"Still enrolled",
"No longer enrolled".
}
}
}

\subsection{Completers}{

An object of class \code{spec_tbl_df}
(inherits from \code{tbl_df}, \code{tbl}, \code{data.frame})
with 18 rows and 5 columns.
\describe{
\item{GR_TYPE}{\verb{<int>} a code that maps to the rest of the columns}
\item{Years}{\verb{<int>} Whether the cohort comes from a 2- or 4-year institution.}
\item{Award Sought}{\verb{<fct>}
Cohorts are defined by their institution and the type of award their students enrolled to earn.
Values include "Certificate", "Associate's", "Bachelor's", or "Any degree" }
\item{Award Level}{\code{chr} the actual credential a person earned, with the same options as above}
\item{Time to Award}{\verb{<fct>} Relative to the normal completion time. Values are "<=100\%", "==125\%", "<=150\%", and "==150\%".}
}
}

}
}
\source{
\url{https://nces.ed.gov/ipeds/datacenter/DataFiles.aspx}
}
\usage{
GLOSSARIES_FOR_GR_TYPES
}
\description{
Mapping details about credentials and success from a complex field in the table of graduation rates.
}
\keyword{datasets}
