% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GLOSSARY_OF_AWARDS}
\alias{GLOSSARY_OF_AWARDS}
\title{Higher Education Award Levels in IPEDS}
\format{
\subsection{\code{GLOSSARY_OF_AWARDS}}{

An object of class \code{spec_tbl_df}
(inherits from \code{tbl_df}, \code{tbl}, \code{data.frame})
with 16 rows and 7 columns.
\describe{
\item{Award Level       }{\verb{<int>} a 4-digit code that captures AWLEVEL and Stage}
\item{AWLEVEL           }{\verb{<int>} essentially the primary key, mostly reflects the order of awards in terms of credits required}
\item{CIP_LEVEL         }{\verb{<chr>} mostly integer codes that are similar to AWLEVEL, except for some short certificates}
\item{Consolidated Level}{\verb{<chr>} short labels derived from documentation for Completers tables. Take with a grain of salt.}
\item{Credential        }{\verb{<fct>} either "Certificate" or "Degree"}
\item{Stage             }{\verb{<fct>} either "Undergraduate" or "Graduate"}
\item{Label             }{\verb{<chr>} a long description of the specific award}
}
}
}
\source{
\url{https://nces.ed.gov/ipeds/datacenter/DataFiles.aspx}
}
\usage{
GLOSSARY_OF_AWARDS
}
\description{
Integer codes for the different kinds of degrees that students may earn.
Every different table in IPEDS has a slightly different way of labeling
credentials like certificates or degrees.
This table attempts to connect them into a single, semi-coherent, system.
}
\keyword{datasets}
