% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GR_PELL_SSL_VARIABLES}
\alias{GR_PELL_SSL_VARIABLES}
\title{Metadata about postsecondary completion by publicly-funded financial assistance}
\format{
\subsection{\code{GR_PELL_SSL_VARIABLES}}{

An object of class \code{spec_tbl_df}
(inherits from \code{tbl_df}, \code{tbl}, \code{data.frame})
with 26 rows and 3 columns.
\describe{
\item{varname}{\code{chr} a seven-character name for each variable}
\item{Assistance}{\code{chr} The type of financial assistance (if any) students received}
\item{Measure}{\code{chr} The quantity reported by the column}
}
}
}
\source{
\url{https://nces.ed.gov/ipeds/datacenter/DataFiles.aspx}
}
\usage{
GR_PELL_SSL_VARIABLES
}
\description{
Metadata about postsecondary completion by publicly-funded financial assistance
}
\keyword{datasets}
