% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GR_VARIABLES}
\alias{GR_VARIABLES}
\title{Definitions of columns found in data that describes postsecondary completers}
\format{
\subsection{\code{GR_VARIABLES}}{

An object of class \code{spec_tbl_df}
(inherits from \code{tbl_df}, \code{tbl}, \code{data.frame})
with 74 rows and 8 columns.
\describe{
\item{varNumber}{\code{int} a code for mapping from raw data to this table.}
\item{varName}{\code{chr} a seven-character name for each variable}
\item{varTitle}{\code{chr} a human-readable name for each variable}
\item{DataType}{\code{chr} "N" for data and "A" for metadata}
\item{format Time}{\code{dbl} either "Cont" (count or continuous) or "Disc" (numeric or character categories)}
\item{Race/Ethnicity}{\code{chr} The OMB racial/ethnic category for people counted by this variable}
\item{Sex}{\code{chr} The OMB gender category for people counted by this variable}
\item{longDescription}{\code{chr} a verbose explanation of the variable}
}
}
}
\source{
\url{https://nces.ed.gov/ipeds/datacenter/DataFiles.aspx}
}
\usage{
GR_VARIABLES
}
\description{
Definitions of columns found in data that describes postsecondary completers
}
\keyword{datasets}
