% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_graduates.R
\name{read_graduates}
\alias{read_graduates}
\title{Load IPEDS data about graduation rates}
\usage{
read_graduates(.dem_file, .aid_file)
}
\arguments{
\item{.dem_file}{\verb{<chr>} the full path to the file with graduation and demographic information.}

\item{.aid_file}{\verb{<chr>} the full path to the file with graduation and financial aid information.}
}
\value{
#' An object of class \code{spec_tbl_df}
(inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}). See Details.
}
\description{
The graduation data come in two tables.
One reports graduation numbers by race/ethnicity/immigration status and sex.
The other reports graduation by type of financial assistance.
Both tables report numbers of students and graduates, and break down by
degree sought and granting institution.
Each file describes one graduation year, so the enrollment year is different
for 2- and 4-year programs.
}
\details{
\subsection{Graduation rates by demographic groupings}{

These data include many different breakdowns, including degree sought, actual
degree earned, several different times to completion and demographic
information such as race/ethnicity, immigration status, and gender.
When you call this function with the "demographic" table option, the output
is a data frame with 15 columns.
\describe{
\item{UNITID            }{\verb{<int>} a seven digit unique identifier for a school}
\item{Index             }{\verb{<int>} the }
\item{Institution Level }{\verb{<chr>}}
\item{Credential Sought }{\verb{<chr>}}
\item{Credential Earned }{\verb{<chr>}}
\item{Relative Time     }{\verb{<chr>}}
\item{Race/Ethnicity    }{\verb{<chr>}}
\item{Sex               }{\verb{<chr>}}
\item{Adjusted Cohort   }{\verb{<int>}}
\item{Completed         }{\verb{<int>}}
\item{Transferred       }{\verb{<int>}}
\item{Full Cohort       }{\verb{<int>}}
\item{Still Enrolled    }{\verb{<int>}}
\item{No Longer Enrolled}{\verb{<int>}}
\item{Excluded          }{\verb{<int>}}
}
}

\subsection{Graduation rates by financial assistance}{

These data all deal with graduation within 150\% of normal completion time.
That means 6 years for bachelor's programs and 3 years for associate's.
When you call this function with the "financial" table option, the output is
a data frame with 9 columns.
\describe{
\item{UNITID             }{\verb{<int>} a seven digit unique identifier for a school}
\item{PSGRTYPE           }{\verb{<int>} a one-digit code describing the type of credential earned}
\item{Assistance         }{\verb{<chr>} the type of assistance that students receive, if any}
\item{Revised Cohort     }{\verb{<int>} an overestimate of the number of students who could have graduated}
\item{Exclusions         }{\verb{<int>} the number of students who should be removed from the denominator}
\item{Adjusted Cohort    }{\verb{<int>} a better estimate of the number of students who could have graduated}
\item{Bachelor's Degrees }{\verb{<int>} the number of students who earned a 4-year degree}
\item{Associate's Degrees}{\verb{<int>} the number of students who earned a degree that took fewer than 4 years}
\item{All Degrees        }{\verb{<int>} the total number of undegraduate degrees earned}
}
}
}
\seealso{
\code{\link[=read_graduates_and_demographics]{read_graduates_and_demographics()}}

\code{\link[=read_graduates_and_assistance]{read_graduates_and_assistance()}}
}
