% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_graduates_and_assistance.R
\name{read_graduates_and_assistance}
\alias{read_graduates_and_assistance}
\title{Read data about graduation and financial assistance}
\usage{
read_graduates_and_assistance(.file)
}
\arguments{
\item{.file}{the full path to the file}
}
\value{
An object,
of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}),
with 9 columns.
\describe{
\item{UNITID       }{\verb{<int>} a seven digit unique identifier for a school}
\item{Years        }{\verb{<int>} Whether the cohort comes from a 2- or 4-year institution.}
\item{Award Sought }{\verb{<fct>}
Cohorts are defined by their institution and the type of award their students enrolled to earn.
Values include "Certificate", "Associate's", "Bachelor's", or "Any degree"
}
\item{Assistance   }{\verb{<chr>} the type of assistance that students receive, if any}
\item{Cohort       }{\verb{<tbl>} information about folks seeking credentials. See Details}
\item{Completers   }{\verb{<tbl>} information about folks earning credentials. See Details}
}
}
\description{
IPEDS tables with names like "GRXXXX_PELL_SSL" contain data about
postsecondary completion and publicly-funded financial assistance.
}
\details{
All of these completion numbers report 150\% graduation rates.
There are two list-columns that contain nested data frames.
Their structures are:
\subsection{Cohort}{

\describe{
\item{Statistic}{\verb{<chr>} the quantity being measured, like cohort size or number of students excluded.}
\item{Count    }{\verb{<chr>} the corresponding number of students. }
}
}

\subsection{Completers}{

\describe{
\item{Award Level}{\verb{<chr>} one of "Associate's Degrees", "Bachelor's Degrees", or "All Degrees".}
\item{Count      }{\verb{<int>} the corresponding number of degrees earned}
}
}
}
\seealso{
\code{\link[=read_graduates]{read_graduates()}} to see why the return value has such a weird structure.
}
