% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_graduates_and_demographics.R
\name{read_graduates_and_demographics}
\alias{read_graduates_and_demographics}
\title{Read data about graduation rates by demographic groupings}
\usage{
read_graduates_and_demographics(.file)
}
\arguments{
\item{.file}{the full path to the file}
}
\value{
An object,
of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}),
with 9 columns.
\describe{
\item{UNITID       }{\verb{<int>} a seven digit unique identifier for a school}
\item{Years        }{\verb{<int>} Whether the cohort comes from a 2- or 4-year institution.}
\item{Award Sought }{\verb{<fct>}
Cohorts are defined by their institution and the type of award their students enrolled to earn.
Values include "Certificate", "Associate's", "Bachelor's", or "Any degree"
}
\item{Race/Ethnicity}{\verb{<chr>} one of the OMB '97 categories}
\item{Sex           }{\verb{<chr>} all, men, or women.}
\item{Cohort       }{\verb{<tbl>} information about folks seeking credentials. See Details}
\item{Completers   }{\verb{<tbl>} information about folks earning credentials. See Details}
}
}
\description{
This table reports numbers of students and graduates, and break down by
degree sought and granting institution.
Each file describes one graduation year, so the enrollment year is different
for 2- and 4-year programs.
These data include many different breakdowns, including degree sought, actual
degree earned, several different times to completion and demographic
information such as race/ethnicity, immigration status, and gender.
When you call this function with the "demographic" table option, the output
}
\details{
The return value has two list-columns that contain nested data frames.
Their structures are:
\subsection{Cohort}{

\describe{
\item{Statistic}{\verb{<chr>} the quantity being measured, like cohort size or number of students excluded.}
\item{Count    }{\verb{<chr>} the corresponding number of students. }
}
}

\subsection{Completers}{

\describe{
\item{Award Level}{\verb{<chr>} one of "Associate's Degrees", "Bachelor's Degrees", or "All Degrees".}
\item{Count      }{\verb{<int>} the corresponding number of degrees earned}
}
}
}
\seealso{
\code{\link[=read_graduates]{read_graduates()}} to see why the return value has such a weird structure.
}
