% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_graduation_rates.R
\name{read_graduation_rates}
\alias{read_graduation_rates}
\title{Load IPEDS data about graduation rates.}
\usage{
read_graduation_rates(.year, .path)
}
\arguments{
\item{.year}{\emph{<int>} the academic year that the information describes}

\item{.path}{\emph{<chr>} the full path to the zip archive that contains the file.}
}
\value{
a tibble of many columns
}
\description{
Load IPEDS data about graduation rates.
}
