% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpack_variable.R
\name{unpack_variable}
\alias{unpack_variable}
\title{Expand a compact coded variable into its multiple meanings in a lookup table}
\usage{
unpack_variable(
  .table_with_variable,
  .variable_definitions,
  .variable_name,
  ...
)
}
\arguments{
\item{.table_with_variable}{\verb{<tbl>} the table where the variable appears without elaboration}

\item{.variable_definitions}{\verb{<tbl>} the table of definitions for the variable}

\item{.variable_name}{\verb{<any>} the name of the variable field, which should be the same in both of the other inputs}

\item{...}{\verb{<chr>} fields from \code{variable_definitions} to keep. Leaving this empty means "use all columns in \code{.variable_definitions}"}
}
\value{
\verb{<tbl>} a data frame with all columns in \code{.table_with_variable} (except \code{.variable_name}) and the \code{...} columns from \code{.variable_definitions}.
}
\description{
Many situations arise where a single variable, like "GRTYPE" in IPEDS,
actually encodes a specific combination of values from several other
variables.
This function is a utility for adding all of those meanings to a table that
originally just has the compact one.
}
