% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_completers.R
\name{wrangle_completers}
\alias{wrangle_completers}
\title{Render IPEDS college completion data into a long, tidy tibble.}
\usage{
wrangle_completers(.conn, .spec)
}
\arguments{
\item{.conn}{\emph{<chr>} a suitable input to \code{\link[readr:read_delim]{readr::read_csv()}}, as from \code{\link[=unzip_completers]{unzip_completers()}}}

\item{.spec}{\emph{<tbl>} one of the elements of \code{COMPLETERS_SPEC}}
}
\value{
\emph{<tbl>} college completion data, aggregated according to \code{.spec}
}
\description{
Render IPEDS college completion data into a long, tidy tibble.
}
\seealso{
\code{\link[=unzip_completers]{unzip_completers()}}

\code{\link[=read_completers]{read_completers()}}
}
