% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AWLEVEL}
\alias{AWLEVEL}
\title{Higher Education Award Levels in IPEDS}
\format{
\subsection{\code{AWLEVEL}}{

An object of class \code{spec_tbl_df}
(inherits from \code{tbl_df}, \code{tbl}, \code{data.frame})
with 16 rows and 2 columns.
\describe{
\item{AWLEVEL}{\emph{<int>} Code value for a degree type}
\item{Award Level}{\emph{<chr>} Human-readable name for a degree type}
}
}
}
\source{
\url{https://nces.ed.gov/ipeds/datacenter/DataFiles.aspx}
}
\usage{
AWLEVEL
}
\description{
Integer codes for the different kinds of degrees that students may earn.
}
\keyword{datasets}
