% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CIP_FAMILIES}
\alias{CIP_FAMILIES}
\title{Very broad categories of subject areas for post-secondary credentials.}
\format{
\subsection{\code{CIP_FAMILIES}}{

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 50 rows and 3 columns.
\subsection{Code}{

A character vector of two-digit numeric codes
}

\subsection{Title}{

A somewhat short description of the subject area. IN ALL CAPS FOR DRAMA
}

\subsection{Definition}{

A somewhat longer description of each subject area. In sentence case.
}

}
}
\source{
\url{https://nces.ed.gov/ipeds/cipcode/resources.aspx}
}
\usage{
CIP_FAMILIES
}
\description{
Very broad categories of subject areas for post-secondary credentials.
}
\keyword{datasets}
