% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CORNELLS}
\alias{CORNELLS}
\title{IPEDS from 2022 for schools with "Cornell" in their names.}
\format{
\subsection{\code{cornells}}{

A list with two elements.
\subsection{Directory}{

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 3 rows and 73 columns.
}

\subsection{Completers}{

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 7 rows and 36 columns.
}

}
}
\source{
\url{https://nces.ed.gov/ipeds/datacenter/DataFiles.aspx}
}
\usage{
CORNELLS
}
\description{
IPEDS from 2022 for schools with "Cornell" in their names.
}
\keyword{datasets}
