% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_outcomes.R
\name{.join_and_drop_keys}
\alias{.join_and_drop_keys}
\title{A common task is to connect to tables by a common foreign key, then drop the key}
\usage{
.join_and_drop_keys(.lhs, .rhs, .keys)
}
\arguments{
\item{.lhs}{one table}

\item{.rhs}{another one to join it to}

\item{.keys}{the column(s) to join them by}
}
\value{
a new table
}
\description{
A common task is to connect to tables by a common foreign key, then drop the key
}
