% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_annual_database.R
\name{fetch_annual_database}
\alias{fetch_annual_database}
\title{Download an Access-formatted database of IPEDs data for one year}
\usage{
fetch_annual_database(.year, .folder_path)
}
\arguments{
\item{.year}{\verb{<int>} the starting year the data describe. Usually July-June.}

\item{.folder_path}{\verb{<chr>} the location to save the downloaded files to.}
}
\description{
Download an Access-formatted database of IPEDs data for one year
}
\seealso{
\code{\link[curl:curl_download]{curl::curl_download()}}

\code{\link[httr:HEAD]{httr::HEAD()}}

\code{\link[httr:status_code]{httr::status_code()}}
}
