% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_annual_database.R
\name{fetch_annual_database}
\alias{fetch_annual_database}
\title{Download an Access-formatted database of IPEDs data for one year}
\usage{
fetch_annual_database(.year, .folder_path)
}
\arguments{
\item{.year}{\if{html}{\out{<int>}} the starting year the data describe. Usually July-June.}

\item{.folder_path}{\if{html}{\out{<str>}} the location to save the downloaded files to.}
}
\value{
\if{html}{\out{<null>}} no return value
}
\description{
Download an Access-formatted database of IPEDs data for one year
}
