% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_yearly_tables.R
\name{find_yearly_tables}
\alias{find_yearly_tables}
\title{Search a path for files with IPEDS-type names}
\usage{
find_yearly_tables(.path, .prefix, .suffix = "", ...)
}
\arguments{
\item{.path}{\verb{<chr>} the path to search for.}

\item{.prefix}{\verb{<chr>} the prefix for this type of table}

\item{.suffix}{\verb{<chr?>} an optional suffix to file names. Defaults to ""}

\item{...}{
  Arguments passed on to \code{\link[base:dir]{base::dir}}
  \describe{
    \item{\code{all.files}}{a logical value.  If \code{FALSE}, only the
    names of visible files are returned (following Unix-style visibility,
    that is files whose name does not start with a dot).  If \code{TRUE},
    all file names will be returned.}
    \item{\code{full.names}}{a logical value.  If \code{TRUE}, the directory
    path is prepended to the file names to give a relative file path.
    If \code{FALSE}, the file names (rather than paths) are returned.}
    \item{\code{recursive}}{logical.  Should the listing recurse into directories?}
    \item{\code{ignore.case}}{logical.  Should pattern-matching be case-insensitive?}
    \item{\code{include.dirs}}{logical.  Should subdirectory names be included in
    recursive listings?  (They always are in non-recursive ones).}
    \item{\code{no..}}{logical.  Should both \code{"."} and \code{".."} be excluded
    also from non-recursive listings?}
    \item{\code{fixed}}{logical.  If \code{TRUE}, treat \code{pattern} as a literal
    string to match rather than a regular expression.}
  }}
}
\value{
\verb{<chr[]>} a vector of file names, with names set to matching years
}
\description{
Files that are downloaded from IPEDS typically have a constant format within
each type of table.
The only difference among files is a year number.
For example, a file for graduation rates from 1998 would be "GR1998.xlsx"
}
