% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GLOSSARIES_FOR_EFFY}
\alias{GLOSSARIES_FOR_EFFY}
\title{Mapping details about enrollment from a complex field in the table of head counts.}
\format{
\subsection{\code{GLOSSARIES_FOR_EFFY}}{

A list of five data frames.
\subsection{EFFYALEV}{

An object of class \code{spec_tbl_df}
(inherits from \code{tbl_df}, \code{tbl}, \code{data.frame})
with 27 rows and 2 columns.

\describe{
\item{EFFYALEV}{\verb{<int>} a code that maps to the rest of the columns}
\item{Label}{\verb{<chr>} some detail about enrollment status for subdividing students}
}
}

\subsection{EFFYLEV}{

An object of class \code{spec_tbl_df}
(inherits from \code{tbl_df}, \code{tbl}, \code{data.frame})
with 4 rows and 2 columns.

\describe{
\item{EFFYLEV}{\verb{<int>} a code that maps to the rest of the columns}
\item{Level of Study}{\verb{<chr>} All, Graduate, Undergraduate, or Detail}
}
}

\subsection{LSTUDY}{

An object of class \code{spec_tbl_df}
(inherits from \code{tbl_df}, \code{tbl}, \code{data.frame})
with 3 rows and 2 columns.

\describe{
\item{LSTUDY}{\verb{<int>} a code that maps to the rest of the columns}
\item{Original Level of Study}{\verb{<chr>} TOtal, Graduate, or Undergraduate}
}
}

\subsection{Population}{

An object of class \code{spec_tbl_df}
(inherits from \code{tbl_df}, \code{tbl}, \code{data.frame})
with 14 rows and 2 columns.

\describe{
\item{EFFPOP}{\verb{<int>} a code that maps to the rest of the columns}
\item{Description}{\verb{<chr>} The population description from the dictionary file}
\item{Population}{\verb{<chr>} A population name similar to those from OMB97}
}
}

\subsection{Sex}{

An object of class \code{spec_tbl_df}
(inherits from \code{tbl_df}, \code{tbl}, \code{data.frame})
with 4 rows and 2 columns.

\describe{
\item{EFFYSEX}{\verb{<chr>} a one-letter code that maps to the rest of the columns}
\item{Sex}{\verb{<chr>} All, Female, Male, or a disastrous mess that attempts to deal with non-binary genders.}
}
}

}
}
\source{
\url{https://nces.ed.gov/ipeds/datacenter/DataFiles.aspx}
}
\usage{
GLOSSARIES_FOR_EFFY
}
\description{
Mapping details about enrollment from a complex field in the table of head counts.
}
\keyword{datasets}
