% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GRTYPE}
\alias{GRTYPE}
\title{Mapping details about credentials and success from a complex field in the table of graduation rates.}
\format{
\subsection{\code{GRTYPE}}{

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df},
\code{tbl}, \code{data.frame}) with 40 rows and 7 columns.
\describe{
\item{GRTYPE}{\if{html}{\out{<int>}} a code that maps to the rest of the columns}
\item{Definition}{\if{html}{\out{<chr>}} the definition from the IPEDs dictionary file.}
\item{Institution Type}{\if{html}{\out{<chr>}} either "4-year" or "2-year"}
\item{Subcohort}{\if{html}{\out{<chr>}} the original degree a person sought: "All," "Bachelor's," "Degree/certificate", or "Other degree/certificate"}
\item{Outcome}{\if{html}{\out{<chr>}} An endpoint such as "Completed," "No longer enrolled," or "Transferred out."}
\item{Credential}{\if{html}{\out{<chr>}} The kind of degree the person ultimately earned.}
\item{\verb{Duration Ratio}}{\if{html}{\out{<dbl>}} The ratio of the time it took to complete the degree to the expected time to complete the degree, e.g. finishing a Bachelor's degree in 6 years counts as 1.5}
}
}
}
\source{
\url{https://nces.ed.gov/ipeds/datacenter/DataFiles.aspx}
}
\usage{
GRTYPE
}
\description{
Mapping details about credentials and success from a complex field in the table of graduation rates.
}
\keyword{datasets}
