% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hercipeds-package.R
\docType{package}
\name{hercipeds-package}
\alias{hercipeds}
\alias{hercipeds-package}
\title{hercipeds: Work with data from the Integrated Postsecondary Education Data System}
\description{
IPEDS contains data in many different tables. The format is well-defined, but not particularly tidy. This package wrangles the data into tidy format and translates the machine-readable field names with their human-readable equivalents.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://higherx4racine.github.io/hercipeds/}
}

}
\author{
\strong{Maintainer}: Ben Taft \email{ben.taft@career2cradle.org}

Authors:
\itemize{
  \item Ben Taft \email{ben.taft@career2cradle.org}
}

Other contributors:
\itemize{
  \item Higher Expectations for Racine County [copyright holder]
}

}
\keyword{internal}
