% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{OMCHRT}
\alias{OMCHRT}
\title{Cohort details for counting 4-, 6-, and 8-year completion dates}
\format{
\subsection{\code{OMCHRT}}{

An object of class \code{spec_tbl_df}
(inherits from \code{tbl_df}, \code{tbl}, \code{data.frame})
with 15 rows and 4 columns.
\describe{
\item{OMCHRT}{\code{int} Code value for cohort type}
\item{Entry Status}{\code{chr} All, First-time, or Not-first-time}
\item{Time Status}{\code{chr} All, Full-time, or Part-time}
\item{Aid Status}{\code{chr} All, Pell recipient, or not Pell recipient}
}
}
}
\source{
\url{https://nces.ed.gov/ipeds/datacenter/DataFiles.aspx}
}
\usage{
OMCHRT
}
\description{
Integer codes for different combinations of first-time, full-time, and financial aid statuses
}
\keyword{datasets}
