% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{OUTCOMES_SPEC}
\alias{OUTCOMES_SPEC}
\title{Layout information about the tables that report 4-, 6-, and 8-year completion rates}
\format{
\subsection{\code{OUTCOMES_SPEC}}{

In addition to the basic ID and name info, lots of fields about location and programming
}
}
\source{
\url{https://nces.ed.gov/ipeds/datacenter/DataFiles.aspx}
}
\usage{
OUTCOMES_SPEC
}
\description{
Layout information about the tables that report 4-, 6-, and 8-year completion rates
}
\keyword{datasets}
