% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cip_codes.R
\name{read_cip_codes}
\alias{read_cip_codes}
\title{Load a table of definitions for IPEDS CIP codes.}
\usage{
read_cip_codes(.path)
}
\arguments{
\item{.path}{\if{html}{\out{<chr>}} the full path to a file downloaded from NCES.}
}
\value{
a \code{\link[tibble:tibble]{tibble::tibble()}} with five variables
\describe{
\item{Family}{\if{html}{\out{<chr>}} a two-digit code for the broadest categories of programs}
\item{Code}{\if{html}{\out{<chr>}} a 2-, 4-, or 6-digit code in ##[.####] format, family first.}
\item{Title}{\if{html}{\out{<chr>}} a short-ish label for the group or program}
\item{Definition}{\if{html}{\out{<chr>}} a longer description of the group or program}
\item{Type}{\if{html}{\out{<chr>}} whether the code has 2, 4, or 6 digits}
}
}
\description{
The Classification of Instructional Programs is a taxonomy of higher
education programs/majors/degrees/etc. It is maintained by the
\href{https://nces.ed.gov/ipeds/cipcode/default.aspx}{National Center for Education Statistics}
}
