% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_completers.R
\name{read_completers}
\alias{read_completers}
\title{Read data about college completers from one zipped file}
\usage{
read_completers(.year, .path = ".", .aggregation = "C")
}
\arguments{
\item{.year}{\if{html}{\out{<int>}} - the academic year the data correspond to}

\item{.path}{\if{html}{\out{<chr>}}, optional - the location of the zipped file, defaults to "."}

\item{.aggregation}{\if{html}{\out{<chr>}}, optional - "A", "B", or "C" (the default), see Notes}
}
\value{
a data frame with approximately 5 columns:
\description{
\item{UNITID}{\if{html}{\out{<int>}} The unique identifier of a single higher education institution}
\item{Population}{\if{html}{\out{<chr>}} A demographic group like "US Residents", "Asian", or "18-24"}
\item{Gender}{\if{html}{\out{<chr>}} Usually "Female" or "Male", but possibly "Another" or "Unknown"}
\item{Completers}{\if{html}{\out{<int>}} The number of people in this population-gender combination}
}
}
\description{
Each file contains the number of students receiving a degree or certificate.
Data cover awards granted between July 1, \code{.year - 1} and June 30, \code{.year}.
}
\details{
\subsection{Aggregation C}{

This version of the table enumerates students by level of award and by
race/ethnicity, gender and age categories. Each file contains multiple
records per institution. Each record is uniquely defined by the variables
IPEDS ID (UNITID) and award level (AWLEVELC ). The 'C' is stripped from
"AWLEVELC" during post-read processing. Each record will contain:
\itemize{
\item the total number of students receiving awards
\item number of  men and women receiving awards
\item number of students receiving awards for all nine race/ethnicity
categories;
\item number of students receiving awards are also available for the following
age groups:
\itemize{
\item under 18
\item 18-24
\item 25-39
\item 40 and above.
}
}

Numbers are not available for combinations of race, gender and age.
Beginning with 2019-20, the certificate of less than 1 year was replaced by
the following 2 categories:
\itemize{
\item certificates of less than 12 weeks
\item certificates of at least 12-weeks but less-than 1-year.
}
}
}
