% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_directory.R
\name{read_directory}
\alias{read_directory}
\title{Identity and overall characteristics of institutions of higher education}
\usage{
read_directory(.year, .path)
}
\arguments{
\item{.year}{\verb{<int>} the academic year that the information describes}

\item{.path}{\verb{<chr>} the full path to the zip archive that contains the file.}
}
\value{
a tibble of many columns
}
\description{
Identity and overall characteristics of institutions of higher education
}
\seealso{
\code{\link[readr:read_delim]{readr::read_csv()}}
}
