% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_effy.R
\name{read_effy}
\alias{read_effy}
\title{Read 12-month headcount data}
\usage{
read_effy(.filename)
}
\arguments{
\item{.filename}{\verb{<chr>} the full path to the file}
}
\value{
a tibble with 6 or 7 columns
\describe{
\item{UNITID}{\verb{<int>} the 6-digit identifier of a postsecondary institution}
\item{EFFYALEV}{\verb{<int?>} indicates a narrow student subpopulation in later years}
\item{EFFYLEV}{\verb{<int>} indicates undegrad, grad, or all students}
\item{LSTUDY}{\verb{<int>} indicates undegrad, grad, or all students}
\item{Population}{\verb{<chr>} one of several race/ethnicity/origin populations}
\item{Sex}{\verb{<chr>} gender identities}
\item{Students}{\verb{<int>} the number of students}
}
}
\description{
Read 12-month headcount data
}
