% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_graduates_and_assistance.R
\name{read_graduates_and_assistance}
\alias{read_graduates_and_assistance}
\title{Load data about postsecondary completion by publicly-funded financial assistance}
\usage{
read_graduates_and_assistance(.file)
}
\arguments{
\item{.file}{the full path to the file}
}
\value{
#' An object of class \code{spec_tbl_df}
(inherits from \code{tbl_df}, \code{tbl}, \code{data.frame})
with 9 columns.
\describe{
\item{UNITID}{\verb{<int>} a seven digit unique identifier for a school}
\item{PSGRTYPE}{\verb{<int>} a one-digit code describing the type of credential earned}
\item{Assistance}{\verb{<chr>} the type of assistance that students receive, if any}
\item{Revised Cohort}{\verb{<int>} an overestimate of the number of students who could have graduated}
\item{Exclusions}{\verb{<int>} the number of students who should be removed from the denominator}
\item{Adjusted Cohort}{\verb{<int>} a better estimate of the number of students who could have graduated}
\item{Bachelor's Degrees}{\verb{<int>} the number of students who earned a 4-year degree}
\item{Associate's Degrees}{\verb{<int>} the number of students who earned a degree that took fewer than 4 years}
\item{All Degrees}{\verb{<int>} the total number of undegraduate degrees earned}
}
}
\description{
All of these completion numbers report 150\% graduation rates.
}
