% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_outcomes.R
\name{read_outcomes}
\alias{read_outcomes}
\title{Load IPEDS data about 4-, 6-, and 8-year completion rates.}
\usage{
read_outcomes(.year, .path)
}
\arguments{
\item{.year}{\if{html}{\out{<int>}} the academic year that the information describes}

\item{.path}{\if{html}{\out{<chr>}} the full path to the zip archive that contains the file.}
}
\value{
a tibble of many columns
}
\description{
Load IPEDS data about 4-, 6-, and 8-year completion rates.
}
