% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unzip_completers.R
\name{unzip_completers}
\alias{unzip_completers}
\title{Read a zipped file of completion data}
\usage{
unzip_completers(.year, .aggregation, .path)
}
\arguments{
\item{.year}{\if{html}{\out{<int>}} the academic year that the completion data come from}

\item{.aggregation}{\if{html}{\out{<chr>}} one of "A", "B", or "C"}

\item{.path}{\if{html}{\out{<chr>}} the full path to the zip file}
}
\value{
a file-like connection to the csv in the zipped archive
}
\description{
\href{https://nces.ed.gov/ipeds/datacenter/DataFiles.aspx}{IPEDS} files come as
giant, zipped csv files. There's no reason to store them on your local
machine in an unzipped format when this function can unzip a file as it reads
its contents into memory.
}
\seealso{
\code{\link[=wrangle_completers]{wrangle_completers()}}
}
